+++ plugins/backend/catalog-backend-module-aws/src/providers/AWSEC2Provider.ts
@@ -39,21 +39,32 @@ export class AWSEC2Provider extends AWSEntityProvider {
       catalogApi?: CatalogApi;
       providerId?: string;
       ownerTag?: string;
+      useTemporaryCredentials?: boolean;
     },
   ) {
     const accountId = config.getString('accountId');
     const roleName = config.getString('roleName');
+    const roleArn = config.getOptionalString('roleArn');
     const externalId = config.getOptionalString('externalId');
     const region = config.getString('region');
 
     return new AWSEC2Provider(
-      { accountId, roleName, externalId, region },
+      { accountId, roleName, roleArn, externalId, region },
       options,
     );
   }
 
   getProviderName(): string {
     return `aws-ec2-provider-${this.accountId}-${this.providerId ?? 0}`;
+  }
+
+  private async getEc2() {
+    const credentials = this.useTemporaryCredentials
+      ? this.getCredentials()
+      : await this.getCredentialsProvider();
+    return this.useTemporaryCredentials
+      ? new EC2({ credentials, region: this.region })
+      : new EC2(credentials);
   }
 
   async run(): Promise<void> {
@@ -65,8 +76,7 @@ export class AWSEC2Provider extends AWSEntityProvider {
     this.logger.info(`Providing ec2 resources from aws: ${this.accountId}`);
     const ec2Resources: ResourceEntity[] = [];
 
-    const credentials = await this.getCredentialsProvider();
-    const ec2 = new EC2(credentials);
+    const ec2 = await this.getEc2();
 
     const defaultAnnotations = this.buildDefaultAnnotations();