+++ plugins/backend/catalog-backend-module-aws/src/providers/AWSOrganizationAccountsProvider.ts
@@ -48,15 +48,17 @@ export class AWSOrganizationAccountsProvider extends AWSEntityProvider {
       catalogApi?: CatalogApi;
       providerId?: string;
       ownerTag?: string;
+      useTemporaryCredentials?: boolean;
     },
   ) {
     const accountId = config.getString('accountId');
     const roleName = config.getString('roleName');
+    const roleArn = config.getOptionalString('roleArn');
     const externalId = config.getOptionalString('externalId');
     const region = config.getString('region');
 
     return new AWSOrganizationAccountsProvider(
-      { accountId, roleName, externalId, region },
+      { accountId, roleName, roleArn, externalId, region },
       options,
     );
   }
@@ -65,6 +67,18 @@ export class AWSOrganizationAccountsProvider extends AWSEntityProvider {
     return `aws-organization-accounts-${this.accountId}-${
       this.providerId ?? 0
     }`;
+  }
+
+  private async getOrganizationsClient() {
+    const credentials = this.useTemporaryCredentials
+      ? this.getCredentials()
+      : await this.getCredentialsProvider();
+    return this.useTemporaryCredentials
+      ? new OrganizationsClient({
+          credentials,
+          region: this.region,
+        })
+      : new OrganizationsClient(credentials);
   }
 
   async run(): Promise<void> {
@@ -78,8 +92,7 @@ export class AWSOrganizationAccountsProvider extends AWSEntityProvider {
     );
     const accountResources: ResourceEntity[] = [];
 
-    const credentials = await this.getCredentialsProvider();
-    const organizationsClient = new OrganizationsClient(credentials);
+    const organizationsClient = await this.getOrganizationsClient();
 
     const defaultAnnotations = this.buildDefaultAnnotations();